pyproject.toml
[project]
name = "trains"
version = "0.1.0"
description = "Real-time train data analysis application for JR-East trains using ODPT API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
